Rename TaskCachingHashesListener to TaskOutputCachingListener
